refactor #138 데이터모델 폴더링
- 데이터모델이 많아, 용이한 구분을 위해 폴더링
